// packages/shared-types/src/enum.ts
export enum ProviderVertical {
    CRM = 'CRM',
    HRIS = 'HRIS',
    ATS = 'ATS',
    Accounting = 'Accounting',
    Ticketing = 'Ticketing',
    MarketingAutomation = 'Marketing Automation',
    FileStorage = 'File Storage',
    Unknown = 'Unknown',
}

export enum CrmProviders {
    ZOHO = 'zoho',
    ZENDESK = 'zendesk',
    HUBSPOT = 'hubspot',
    PIPEDRIVE = 'pipedrive',
    FRESHSALES = 'freshsales',
}
  
export enum AccountingProviders {
    PENNYLANE = 'pennylane',
    FRESHBOOKS = 'freshbooks',
    CLEARBOOKS = 'clearbooks',
    FREEAGENT = 'freeagent',
    SAGE = 'sage',
}
  